Update LDAP_TIME.SQLRPGLE
Changed the date edit mask to include a dash separating the date from the time component. This is more compatible with the IBM i TIMESTAMP to CHAR results users are experiencing. Also changed the EPOC date format and date/time literal to be consistent with that of the new TIMESTAMP format.